change namespace for SoapDomainClientFactory
* move some files for more logical structure for some of the endpoint related classes
