+++ OpenRiaServices.DomainServices.Client.Web/Framework/Web/Behaviors/SoapEndpointBehavior.QueryInspector.cs
@@ -1,0 +1,110 @@
+﻿using System.Collections.Generic;
+using System.Linq;
+using System.Net;
+using System.ServiceModel;
+using System.ServiceModel.Channels;
+using System.ServiceModel.Dispatcher;
+
+namespace OpenRiaServices.DomainServices.Client.Web.Behaviors
+{
+    partial class SoapEndpointBehavior
+    {
+        sealed class QueryInspector : IClientMessageInspector
+        {
+            private const string IncludeTotalCountPropertyName = "DomainServiceIncludeTotalCount";
+            private const string QueryPropertyName = "DomainServiceQuery";
+
+            private readonly WcfDomainClientFactory _factory;
+
+            public QueryInspector(WcfDomainClientFactory factory)
+            {
+                _factory = factory;
+            }
+
+            private static string RewriteAction(string action)
+            {
+                // This might be better of by modifying the 
+                // endpoint during setup via behaviour instead
+                return action.Insert(action.LastIndexOf("/"), "soap");
+            }
+
+            object IClientMessageInspector.BeforeSendRequest(ref Message request, IClientChannel channel)
+            {
+                request.Headers.Action = RewriteAction(request.Headers.Action);
+
+                // Add Query Options if any are set
+                var messageProperties = OperationContext.Current?.OutgoingMessageProperties;
+                if (messageProperties != null)
+                {
+                    messageProperties.TryGetValue(QueryPropertyName, out object queryProperty);
+                    messageProperties.TryGetValue(IncludeTotalCountPropertyName, out object includeTotalCountProperty);
+
+                    // Add Query Options header if any options were specified
+                    if (queryProperty != null || includeTotalCountProperty != null)
+                    {
+                        var queryParts = (queryProperty != null) ? QuerySerializer.Serialize((IQueryable)queryProperty) : null;
+                        var includeTotalCount = (bool?)includeTotalCountProperty;
+
+                        var header = new QueryOptionsHeader(queryParts, includeTotalCount == true);
+                        request.Headers.Add(header);
+                    }
+                }
+
+                return null;
+            }
+
+            void IClientMessageInspector.AfterReceiveReply(ref Message reply, object correlationState)
+            {
+                // Method intentionally left empty.
+            }
+
+            /// <summary>
+            /// Class represponsible for serialization of query options as a SOAP header
+            /// </summary>
+            class QueryOptionsHeader : MessageHeader
+            {
+                private const string QueryOptionElementName = "QueryOption";
+                private const string QueryNameAttribute = "Name";
+                private const string QueryValueAttribute = "Value";
+                private const string QueryIncludeTotalCountOption = "includeTotalCount";
+                private const string QueryHeaderName = "DomainServiceQuery";
+                private const string QueryHeaderNamespace = "DomainServices";
+
+                private readonly IEnumerable<ServiceQueryPart> _queryParts;
+                private readonly bool _includeTotalCount;
+
+                public QueryOptionsHeader(IEnumerable<ServiceQueryPart> queryParts, bool includeTotalCount)
+                {
+                    _queryParts = queryParts;
+                    _includeTotalCount = includeTotalCount;
+                }
+
+                public override string Name => QueryHeaderName;
+
+                public override string Namespace => QueryHeaderNamespace;
+
+                protected override void OnWriteHeaderContents(System.Xml.XmlDictionaryWriter writer, MessageVersion messageVersion)
+                {
+                    if (_queryParts != null)
+                    {
+                        foreach (var part in _queryParts)
+                        {
+                            writer.WriteStartElement(QueryOptionElementName);
+                            writer.WriteAttributeString(QueryNameAttribute, part.QueryOperator);
+                            writer.WriteAttributeString(QueryValueAttribute, part.Expression);
+                            writer.WriteEndElement();
+                        }
+                    }
+
+                    if (_includeTotalCount)
+                    {
+                        writer.WriteStartElement(QueryOptionElementName);
+                        writer.WriteAttributeString(QueryNameAttribute, QueryIncludeTotalCountOption);
+                        writer.WriteAttributeString(QueryValueAttribute, "true");
+                        writer.WriteEndElement();
+                    }
+                }
+            }
+        }
+    }
+}
